Move EditPool redux state to it's own slice
