+++ portal-ui/src/screens/Console/Tenants/TenantDetails/Pools/EditPool/types.ts
@@ -1,0 +1,53 @@
+// This file is part of MinIO Console Server
+// Copyright (c) 2022 MinIO, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import { ITolerationModel } from "../../../../../../common/types";
+import { Opts } from "../../../ListTenants/utils";
+import {
+  IPoolConfiguration,
+  ITenantAffinity,
+  LabelKeyPair,
+} from "../../../types";
+
+export interface IEditPoolSetup {
+  numberOfNodes: number;
+  volumeSize: number;
+  volumesPerServer: number;
+  storageClass: string;
+}
+
+export interface IEditPoolFields {
+  setup: IEditPoolSetup;
+  affinity: ITenantAffinity;
+  configuration: IPoolConfiguration;
+  tolerations: ITolerationModel[];
+  nodeSelectorPairs: LabelKeyPair[];
+}
+
+export interface IEditPool {
+  editPoolLoading: boolean;
+  validPages: string[];
+  storageClasses: Opts[];
+  limitSize: any;
+  fields: IEditPoolFields;
+  editSending: boolean;
+}
+
+export interface PageFieldValue {
+  page: keyof IEditPoolFields;
+  field: string;
+  value: any;
+}
