+++ portal-ui/src/screens/Console/Tenants/tenantsSlice.ts
@@ -15,12 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import {
-  IAddPoolFields,
-  IEditPoolFields,
-  ITenantState,
-  LabelKeyPair,
-} from "./types";
+import { IAddPoolFields, ITenantState, LabelKeyPair } from "./types";
 import {
   ITolerationEffect,
   ITolerationModel,
@@ -29,7 +24,7 @@ import {
 import get from "lodash/get";
 import { has } from "lodash";
 import { Opts } from "./ListTenants/utils";
-import { IPool, ITenant } from "./ListTenants/types";
+import { ITenant } from "./ListTenants/types";
 
 export interface FileValue {
   fileName: string;
@@ -49,12 +44,6 @@ export interface CertificateFile {
   value: string;
 }
 
-export interface PageFieldValue {
-  page: keyof IEditPoolFields;
-  field: string;
-  value: any;
-}
-
 const initialState: ITenantState = {
   tenantDetails: {
     currentTenant: "",
@@ -68,44 +57,6 @@ const initialState: ITenantState = {
   addPool: {
     addPoolLoading: false,
     validPages: ["affinity", "configure"],
-    storageClasses: [],
-    limitSize: {},
-    fields: {
-      setup: {
-        numberOfNodes: 0,
-        storageClass: "",
-        volumeSize: 0,
-        volumesPerServer: 0,
-      },
-      affinity: {
-        nodeSelectorLabels: "",
-        podAffinity: "default",
-        withPodAntiAffinity: true,
-      },
-      configuration: {
-        securityContextEnabled: false,
-        securityContext: {
-          runAsUser: "1000",
-          runAsGroup: "1000",
-          fsGroup: "1000",
-          runAsNonRoot: true,
-        },
-      },
-      nodeSelectorPairs: [{ key: "", value: "" }],
-      tolerations: [
-        {
-          key: "",
-          tolerationSeconds: { seconds: 0 },
-          value: "",
-          effect: ITolerationEffect.NoSchedule,
-          operator: ITolerationOperator.Equal,
-        },
-      ],
-    },
-  },
-  editPool: {
-    editPoolLoading: false,
-    validPages: ["setup", "affinity", "configure"],
     storageClasses: [],
     limitSize: {},
     fields: {
@@ -295,226 +246,6 @@ export const tenantSlice = createSlice({
     setOpenPoolDetails: (state, action: PayloadAction<boolean>) => {
       state.tenantDetails.poolDetailsOpen = action.payload;
     },
-    setInitialPoolDetails: (state, action: PayloadAction<IPool>) => {
-      let podAffinity: "default" | "nodeSelector" | "none" = "none";
-      let withPodAntiAffinity = false;
-      let nodeSelectorLabels = "";
-      let tolerations: ITolerationModel[] = [
-        {
-          key: "",
-          tolerationSeconds: { seconds: 0 },
-          value: "",
-          effect: ITolerationEffect.NoSchedule,
-          operator: ITolerationOperator.Equal,
-        },
-      ];
-      let nodeSelectorPairs: LabelKeyPair[] = [{ key: "", value: "" }];
-
-      if (action.payload.affinity?.nodeAffinity) {
-        podAffinity = "nodeSelector";
-        if (action.payload.affinity?.podAntiAffinity) {
-          withPodAntiAffinity = true;
-        }
-      } else if (action.payload.affinity?.podAntiAffinity) {
-        podAffinity = "default";
-      }
-
-      if (action.payload.affinity?.nodeAffinity) {
-        let labelItems: string[] = [];
-        nodeSelectorPairs = [];
-
-        action.payload.affinity.nodeAffinity.requiredDuringSchedulingIgnoredDuringExecution.nodeSelectorTerms.forEach(
-          (labels) => {
-            labels.matchExpressions.forEach((exp) => {
-              labelItems.push(`${exp.key}=${exp.values.join(",")}`);
-              nodeSelectorPairs.push({
-                key: exp.key,
-                value: exp.values.join(", "),
-              });
-            });
-          }
-        );
-        nodeSelectorLabels = labelItems.join("&");
-      }
-
-      let securityContextOption = false;
-
-      if (action.payload.securityContext) {
-        securityContextOption =
-          !!action.payload.securityContext.runAsUser ||
-          !!action.payload.securityContext.runAsGroup ||
-          !!action.payload.securityContext.fsGroup;
-      }
-
-      if (action.payload.tolerations) {
-        tolerations = action.payload.tolerations?.map((toleration) => {
-          const tolerationItem: ITolerationModel = {
-            key: toleration.key,
-            tolerationSeconds: toleration.tolerationSeconds,
-            value: toleration.value,
-            effect: toleration.effect,
-            operator: toleration.operator,
-          };
-          return tolerationItem;
-        });
-      }
-
-      const volSizeVars = action.payload.volume_configuration.size / 1073741824;
-
-      const newPoolInfoFields: IEditPoolFields = {
-        setup: {
-          numberOfNodes: action.payload.servers,
-          storageClass: action.payload.volume_configuration.storage_class_name,
-          volumeSize: volSizeVars,
-          volumesPerServer: action.payload.volumes_per_server,
-        },
-        configuration: {
-          securityContextEnabled: securityContextOption,
-          securityContext: {
-            runAsUser: action.payload.securityContext?.runAsUser || "",
-            runAsGroup: action.payload.securityContext?.runAsGroup || "",
-            fsGroup: action.payload.securityContext?.fsGroup || "",
-            runAsNonRoot: !!action.payload.securityContext?.runAsNonRoot,
-          },
-        },
-        affinity: {
-          podAffinity,
-          withPodAntiAffinity,
-          nodeSelectorLabels,
-        },
-        tolerations,
-        nodeSelectorPairs,
-      };
-
-      state.editPool.fields = {
-        ...state.editPool.fields,
-        ...newPoolInfoFields,
-      };
-    },
-    setEditPoolLoading: (state, action: PayloadAction<boolean>) => {
-      state.editPool.editPoolLoading = action.payload;
-    },
-    setEditPoolField: (state, action: PayloadAction<PageFieldValue>) => {
-      if (
-        has(
-          state.editPool.fields,
-          `${action.payload.page}.${action.payload.field}`
-        )
-      ) {
-        const originPageNameItems = get(
-          state.editPool.fields,
-          `${action.payload.page}`,
-          {}
-        );
-
-        let newValue: any = {};
-        newValue[action.payload.field] = action.payload.value;
-
-        const joinValue = { ...originPageNameItems, ...newValue };
-
-        state.editPool.fields[action.payload.page] = { ...joinValue };
-      }
-    },
-    isEditPoolPageValid: (
-      state,
-      action: PayloadAction<{
-        page: string;
-        status: boolean;
-      }>
-    ) => {
-      const edPoolPV = [...state.editPool.validPages];
-
-      if (action.payload.status) {
-        if (!edPoolPV.includes(action.payload.page)) {
-          edPoolPV.push(action.payload.page);
-
-          state.editPool.validPages = [...edPoolPV];
-        }
-      } else {
-        const newSetOfPages = edPoolPV.filter(
-          (elm) => elm !== action.payload.page
-        );
-
-        state.editPool.validPages = [...newSetOfPages];
-      }
-    },
-    setEditPoolStorageClasses: (state, action: PayloadAction<Opts[]>) => {
-      state.editPool.storageClasses = action.payload;
-    },
-    setEditPoolTolerationInfo: (
-      state,
-      action: PayloadAction<{
-        index: number;
-        tolerationValue: ITolerationModel;
-      }>
-    ) => {
-      const editPoolTolerationValue = [...state.editPool.fields.tolerations];
-
-      editPoolTolerationValue[action.payload.index] =
-        action.payload.tolerationValue;
-      state.editPool.fields.tolerations = editPoolTolerationValue;
-    },
-    addNewEditPoolToleration: (state) => {
-      state.editPool.fields.tolerations.push({
-        key: "",
-        tolerationSeconds: { seconds: 0 },
-        value: "",
-        effect: ITolerationEffect.NoSchedule,
-        operator: ITolerationOperator.Equal,
-      });
-    },
-    removeEditPoolToleration: (state, action: PayloadAction<number>) => {
-      state.editPool.fields.tolerations =
-        state.editPool.fields.tolerations.filter(
-          (_, index) => index !== action.payload
-        );
-    },
-    setEditPoolKeyValuePairs: (
-      state,
-      action: PayloadAction<LabelKeyPair[]>
-    ) => {
-      state.editPool.fields.nodeSelectorPairs = action.payload;
-    },
-    resetEditPoolForm: (state) => {
-      state.editPool = {
-        editPoolLoading: false,
-        validPages: ["setup", "affinity", "configure"],
-        storageClasses: [],
-        limitSize: {},
-        fields: {
-          setup: {
-            numberOfNodes: 0,
-            storageClass: "",
-            volumeSize: 0,
-            volumesPerServer: 0,
-          },
-          affinity: {
-            nodeSelectorLabels: "",
-            podAffinity: "default",
-            withPodAntiAffinity: true,
-          },
-          configuration: {
-            securityContextEnabled: false,
-            securityContext: {
-              runAsUser: "1000",
-              runAsGroup: "1000",
-              fsGroup: "1000",
-              runAsNonRoot: true,
-            },
-          },
-          nodeSelectorPairs: [{ key: "", value: "" }],
-          tolerations: [
-            {
-              key: "",
-              tolerationSeconds: { seconds: 0 },
-              value: "",
-              effect: ITolerationEffect.NoSchedule,
-              operator: ITolerationOperator.Equal,
-            },
-          ],
-        },
-      };
-    },
   },
 });
 
@@ -535,16 +266,6 @@ export const {
   setPoolKeyValuePairs,
   setSelectedPool,
   setOpenPoolDetails,
-  setInitialPoolDetails,
-  setEditPoolLoading,
-  resetEditPoolForm,
-  setEditPoolField,
-  isEditPoolPageValid,
-  setEditPoolStorageClasses,
-  setEditPoolTolerationInfo,
-  addNewEditPoolToleration,
-  removeEditPoolToleration,
-  setEditPoolKeyValuePairs,
 } = tenantSlice.actions;
 
 export default tenantSlice.reducer;
